Add minBalance to RebasingState to prevent underflow reverts
The rebasing balance of an account was assumed to always exceed the raw ERC20 balance of the same account (the checked subtractions will revert due to underflow if this is not the case). However, this property does not necessarily hold in the case that the account has entered rebasing and the share price has not yet adjusted upwards due to a distribution.

By introducing the minBalance into the RebasingState, user balances are no longer rounded down when they enter rebase, so they can do token transfers without reverting underflows.