Remove LendingTerm.GAUGE_CAP_TOLERANCE()
The majoration of 1.2x on ideal gauge allocation was added to avoid situations where 2 Lending Terms are exactly 50% of the votes each and each issued exactly 50% of the CREDIT totalSupply. This would result in new borrows being stuck on all terms.

However, since we are computing the gauge allocation on the *post-borrow* supply (credit.totalSupply() + borrowAmount), there is already a tolerance of 2x applied.