Add 'borrower' param to LendingTerm.borrow
This will allow helper contracts to open loans on behalf of another address. During borrow, the caller provides collateral & receives the CREDIT debt. If a caller borrows on behalf on another address, the collateral will be returned to that other address in repay() and onBid(). The caller address can choose to forward the CREDIT debt or not after opening the loan.

This opens the possibility to write helper contracts that :
- Can do all normal-flow actions with signed messages (borrow, repay, addCollateral, partialRepay)
- Borrow and redeem the borrowed amount for USDC in the PSM. Mint from USDC in the PSM & repay atomically